backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
networkx==3.3
google-generativeai==0.8.0
firebase-admin==6.5.0
python-dotenv==1.0.1
httpx==0.27.0
xgboost==2.0.3
scikit-learn==1.4.1.post1
